% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUDPC_S}
\alias{AUDPC_S}
\title{A Data Frame of Area Under the Disease Progress Curve and Stairs for Sheath Blight Observations}
\format{
An object of class \code{data.frame} with 88 rows and 21 columns.
}
\usage{
AUDPC_S
}
\description{
A dataset containing the alternate wetting and drying study data observations
and calculated Area Under the Disease Progress Curve (AUDPC) and Stairs
(AUDPS)
}
\keyword{datasets}
