% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUDPS}
\alias{AUDPS}
\title{A Data Frame of Area Under the Disease Progress Stairs for Sheath Blight Observations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 184 rows and 11 columns.
}
\usage{
AUDPS
}
\description{
A dataset containing the alternate wetting and drying study data observations
and calculated Area Under the Disease Progress Stairs (AUDPS)
}
\keyword{datasets}
