% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_table.R
\name{mcmc_table}
\alias{mcmc_table}
\title{MCMCglmm results table}
\usage{
mcmc_table(x)
}
\arguments{
\item{x}{An MCMCglmm model}
}
\description{
Generate a table of MCMCglmm results
}
\examples{
\dontrun{
mcmc_table(MCMC_model)
}

}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
