% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N_rates}
\alias{N_rates}
\title{A Data Frame of Nitrogen Rates and Split Application Timings}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\usage{
N_rates
}
\description{
A dataset containing the N rates and time of application split organised
into a tidy data frame.
}
\keyword{datasets}
