% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates_paper.R
\name{plot_leaf_estimates_paper}
\alias{plot_leaf_estimates_paper}
\title{Plot MCMCglmm leaf estimates for publication}
\usage{
plot_leaf_estimates_paper(x, main_title = "")
}
\arguments{
\item{x}{A \code{\link[coda]{mcmc.list}} object}

\item{main_title}{Main title for the plot}
}
\description{
Plot estimates of posterior distributions and credible
 intervals of \pkg{MCMCglmm} outputs sans intercept for publication figure
}
\examples{
\dontrun{
# Plot posterior means and credible intervals
plot_estimates(m1)
}

}
\author{
Titus von der Malsburg, \email{malsburg@uni-potsdam.de}
}
