% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_mcmc_glmm}
\alias{plot_mcmc_glmm}
\alias{plot_diagnostic_lines}
\alias{plot_replicate_posteriors}
\alias{plot_treatment_posteriors}
\alias{plot_joint_random_error_dist}
\title{MCMCglmm diagnostic plots}
\usage{
plot_diagnostic_lines(d, x, title)

plot_replicate_posteriors(d, title)

plot_treatment_posteriors(d, title)

plot_joint_random_error_dist(d, title)
}
\arguments{
\item{d}{A melted data.frame of MCMCglmm diagnostic outputs}

\item{x}{Dummy x-axis for diagnostic lines plot}

\item{title}{Title to be used in resulting graph output}
}
\description{
Plot diagnostic graphs from MCMCglmm outputs
}
\section{Functions}{

\itemize{
 \item \code{plot_diagnostic_lines()} plots diagnostic lines from MCMCglmm
 output
 \item \code{plot_replicate_posteriors()} plots a violin plot of posteriors
 for experimental plot replicates
 \item \code{plot_treatment_posteriors()} plots a violin plot of posteriors
 for experimental plot treatments
 \item \code{plot_joint_random_error_dist()} plots joint distribution
}
}

\examples{
\dontrun{
# Plot diagnostic lines
plot_diagnostic_lines(d = reps, x = x, title = "Diagnostic Lines Plot")

# Plot replicate posteriors
plot_replicate_posteriors(d = reps, title = "Replicate Posteriors")

# Plot treatment posteriors
plot_treatment_posteriors(d =  trts, title = "Treatment Posteriors")

# Plot joint random error
plot_joint_random_error_dist(d = rdf, title = ("Random Error Distribution"))
}

}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
