% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_tiller_estimates}
\alias{plot_tiller_estimates}
\title{Plot MCMCglmm tiller estimates}
\usage{
plot_tiller_estimates(x)
}
\arguments{
\item{x}{An\code{\link[coda]{mcmc.list}} object}
}
\description{
Plot estimates of posterior distributions and credible
 intervals of \pkg{MCMCglmm} outputs
}
\examples{
\dontrun{
# Plot posterior means and credible intervals
plot_tiller_estimates(m1)
}

}
\author{
Titus von der Malsburg, \email{malsburg@uni-potsdam.de}
}
