% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAW_data}
\alias{RAW_data}
\title{A Tidy Data frame of Raw Data for Sheath Blight Observations}
\format{
An object of class \code{data.frame} with 79488 rows and 23 columns.
}
\usage{
RAW_data
}
\description{
A dataset containing the alternate wetting and drying study data observations
organised into a tidy data frame.
}
\keyword{datasets}
