% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.awd.pests.R
\docType{package}
\name{rice.awd.pests-package}
\alias{rice.awd.pests-package}
\title{Reproducible Research Compendium for Analysing Effects of Water Management and Nitrogen on Rice Sheath Blight and Insect Populations}
\description{
Provides a research compendium of analysis associated with experiments
conducted at the International Rice Research Institute (IRRI), Los Baños,
Calabarzon, Philippines in the dry seasons of 2015 and 2016. The experiments
were designed to test the effects of alternate wetting and drying irrigation
management on rice sheath blight disease. The plots were split plot with an
irrigation main plot and nitrogen sub-plot treatments.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{\url{https://github.com/phytopathology/rice_awd_pests}}
\item{Report bugs at
 \url{https://github.com/phytopathology/rice_awd_pests/issues}}
\item{Research compendium website
 \url{https://phytopathology.github.io/rice_awd_pests/}}
}
}
\author{
Adam H Sparks and Nancy P Castilla and Michael Noel
}
