% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.awd.shb-package.R
\docType{package}
\name{rice.awd.shb-package}
\alias{rice.awd.shb}
\alias{rice.awd.shb-package}
\title{rice.awd.shb: Reproducible Research Compendium for Analysing Effects of Water Management and Nitrogen on Rice Sheath Blight}
\description{
Provides a research compendium of analysis associated with
    experiments conducted at the International Rice Research Institute (IRRI),
    Los Baños, Calabarzon, Philippines in the dry seasons of 2015 and 2016. The
    experiments were designed to test the effects of alternate wetting and
    drying irrigation management on rice sheath blight disease. The plots were
    split plot with an irrigation main plot and nitrogen sub-plot treatments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/rice-awd-shb}
  \item Report bugs at \url{https://github.com/adamhsparks/rice-awd-shb/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Nancy P. Castilla \email{n.castilla@irri.org}
  \item Bjoern Ole Sander \email{b.sander@irri.org} (\href{https://orcid.org/0000-0002-7967-6147}{ORCID})
  \item Michael Noel
}

Other contributors:
\itemize{
  \item Global Rice Science Partnership (GRiSP) [funder]
}

}
\keyword{internal}
