% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates_paper.R
\name{plot_estimates_paper}
\alias{plot_estimates_paper}
\title{Plot MCMCglmm estimates for publication}
\usage{
plot_estimates_paper(x)
}
\arguments{
\item{x}{An\code{\link[coda]{mcmc.list}} object}
}
\description{
Plot estimates of posterior distributions and credible
 intervals of MCMCglmm outputs sans intercept for publication figure
}
\examples{
\dontrun{
# Plot posterior means and credible intervals
plot_estimates(m1)
}

}
\author{
Titus von der Malsburg, \email{malsburg@uni-potsdam.de}
}
